refactor rover-cli build script; remove fallback assets path
