+++ rover-cli/build_script/assets/runtime_index.html
@@ -1,0 +1,12 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Rover Web</title>
+</head>
+<body>
+  <div id="app"></div>
+  <script type="module" src="./loader.js"></script>
+</body>
+</html>
