+++ rover-cli/build_script/embedded.rs
@@ -1,0 +1,7 @@
+pub fn runtime_index_html() -> &'static str {
+    include_str!("assets/runtime_index.html")
+}
+
+pub fn runtime_loader_js() -> &'static str {
+    include_str!("assets/runtime_loader.js")
+}
